# Conflicts:
#	build.gradle
#	gradle.properties
#	gradle/wrapper/gradle-wrapper.jar
#	gradle/wrapper/gradle-wrapper.properties
#	src/main/java/com/jozufozu/flywheel/Flywheel.java
#	src/main/java/com/jozufozu/flywheel/backend/ShaderSources.java
#	src/main/java/com/jozufozu/flywheel/backend/instancing/InstanceMaterial.java
#	src/main/java/com/jozufozu/flywheel/backend/instancing/InstancedRenderDispatcher.java
#	src/main/java/com/jozufozu/flywheel/backend/instancing/MaterialManager.java
#	src/main/java/com/jozufozu/flywheel/config/BooleanConfig.java
#	src/main/java/com/jozufozu/flywheel/config/BooleanConfigCommand.java
#	src/main/java/com/jozufozu/flywheel/config/FlwCommands.java
#	src/main/java/com/jozufozu/flywheel/config/FlwConfig.java
#	src/main/java/com/jozufozu/flywheel/core/AtlasStitcher.java
#	src/main/java/com/jozufozu/flywheel/core/crumbling/CrumblingMaterialManager.java
#	src/main/java/com/jozufozu/flywheel/core/crumbling/CrumblingRenderer.java
#	src/main/java/com/jozufozu/flywheel/event/BeginFrameEvent.java
#	src/main/java/com/jozufozu/flywheel/event/EntityWorldHandler.java
#	src/main/java/com/jozufozu/flywheel/event/ForgeEvents.java
#	src/main/java/com/jozufozu/flywheel/event/ReloadRenderersEvent.java
#	src/main/java/com/jozufozu/flywheel/event/RenderLayerEvent.java
#	src/main/java/com/jozufozu/flywheel/mixin/RenderHooksMixin.java
#	src/main/java/com/jozufozu/flywheel/mixin/TileWorldHookMixin.java
#	src/main/java/com/jozufozu/flywheel/util/BufferBuilderReader.java
#	src/main/java/com/jozufozu/flywheel/util/RenderUtil.java
#	src/main/resources/META-INF/mods.toml
#	src/main/resources/flywheel.mixins.json

+++ src/main/java/com/jozufozu/flywheel/backend/instancing/Instancer.java
@@ -1,6 +1,5 @@
 package com.jozufozu.flywheel.backend.instancing;
 
-
 import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.function.Supplier;
@@ -11,16 +10,34 @@ import com.jozufozu.flywheel.backend.gl.attrib.VertexFormat;
 import com.jozufozu.flywheel.backend.gl.buffer.GlBuffer;
 import com.jozufozu.flywheel.backend.gl.buffer.GlBufferType;
 import com.jozufozu.flywheel.backend.gl.buffer.MappedBuffer;
-import com.jozufozu.flywheel.backend.model.BufferedModel;
+import com.jozufozu.flywheel.backend.material.MaterialSpec;
+import com.jozufozu.flywheel.backend.model.IBufferedModel;
+import com.jozufozu.flywheel.backend.model.IndexedModel;
+import com.jozufozu.flywheel.core.model.IModel;
 import com.jozufozu.flywheel.util.AttribUtil;
 
-import net.minecraft.util.math.vector.Vector3i;
-
+/**
+ * An instancer is how you interact with an instanced model.
+ * <p>
+ *     Instanced models can have many copies, and on most systems it's very fast to draw all of the copies at once.
+ *     There is no limit to how many copies an instanced model can have.
+ *     Each copy is represented by an InstanceData object.
+ * </p>
+ * <p>
+ *     When you call {@link #createInstance()} you are given an InstanceData object that you can manipulate however
+ *     you want. The changes you make to the InstanceData object are automatically made visible, and persistent.
+ *     Changing the position of your InstanceData object every frame means that that copy of the model will be in a
+ *     different position in the world each frame. Setting the position of your InstanceData once and not touching it
+ *     again means that your model will be in the same position in the world every frame. This persistence is useful
+ *     because it means the properties of your model don't have to be re-evaluated every frame.
+ * </p>
+ *
+ * @param <D> the data that represents a copy of the instanced model.
+ */
 public class Instancer<D extends InstanceData> {
 
-	public final Supplier<Vector3i> originCoordinate;
-
-	protected final BufferedModel model;
+	protected final Supplier<IModel> gen;
+	protected IBufferedModel model;
 
 	protected final VertexFormat instanceFormat;
 	protected final IInstanceFactory<D> factory;
@@ -29,17 +46,54 @@ public class Instancer<D extends InstanceData> {
 	protected int glBufferSize = -1;
 	protected int glInstanceCount = 0;
 	private boolean deleted;
+	private boolean initialized;
 
 	protected final ArrayList<D> data = new ArrayList<>();
 
 	boolean anyToRemove;
 	boolean anyToUpdate;
 
-	public Instancer(BufferedModel model, Supplier<Vector3i> originCoordinate, MaterialSpec<D> spec) {
-		this.model = model;
+	public Instancer(Supplier<IModel> model, MaterialSpec<D> spec) {
+		this.gen = model;
 		this.factory = spec.getInstanceFactory();
 		this.instanceFormat = spec.getInstanceFormat();
-		this.originCoordinate = originCoordinate;
+	}
+
+	/**
+	 * @return a handle to a new copy of this model.
+	 */
+	public D createInstance() {
+		return _add(factory.create(this));
+	}
+
+	/**
+	 * Copy a data from another Instancer to this.
+	 *
+	 * This has the effect of swapping out one model for another.
+	 * @param inOther the data associated with a different model.
+	 */
+	public void stealInstance(D inOther) {
+		if (inOther.owner == this) return;
+
+		inOther.owner.anyToRemove = true;
+		_add(inOther);
+	}
+
+	public void render() {
+		if (!isInitialized()) init();
+		if (deleted) return;
+
+		vao.bind();
+		renderSetup();
+
+		if (glInstanceCount > 0) model.drawInstances(glInstanceCount);
+
+		vao.unbind();
+	}
+
+	private void init() {
+		model = new IndexedModel(gen.get());
+		initialized = true;
 
 		if (model.getVertexCount() <= 0)
 			throw new IllegalArgumentException("Refusing to instance a model with no vertices.");
@@ -59,27 +113,11 @@ public class Instancer<D extends InstanceData> {
 		model.clearState();
 	}
 
-	public void render() {
-		if (deleted) return;
-
-		vao.bind();
-		renderSetup();
-
-		if (glInstanceCount > 0) model.drawInstances(glInstanceCount);
-
-		vao.unbind();
+	public boolean isInitialized() {
+		return initialized;
 	}
 
-	public D createInstance() {
-		D instanceData = factory.create(this);
-		instanceData.dirty = true;
-		anyToUpdate = true;
-		data.add(instanceData);
-
-		return instanceData;
-	}
-
-	public boolean empty() {
+	public boolean isEmpty() {
 		return !anyToUpdate && !anyToRemove && glInstanceCount == 0;
 	}
 
@@ -92,17 +130,31 @@ public class Instancer<D extends InstanceData> {
 	}
 
 	/**
-	 * Free acquired resources. Attempting to use this after calling delete is undefined behavior.
+	 * Free acquired resources. All other Instancer methods are undefined behavior after calling delete.
 	 */
 	public void delete() {
 		if (deleted) return;
 
 		deleted = true;
 
-		model.delete();
+		if (isInitialized()) {
+			model.delete();
 
-		instanceVBO.delete();
-		vao.delete();
+			instanceVBO.delete();
+			vao.delete();
+		}
+	}
+
+	private D _add(D instanceData) {
+		instanceData.owner = this;
+
+		instanceData.dirty = true;
+		anyToUpdate = true;
+		synchronized (data) {
+			data.add(instanceData);
+		}
+
+		return instanceData;
 	}
 
 	protected void renderSetup() {
@@ -215,7 +267,7 @@ public class Instancer<D extends InstanceData> {
 		final BitSet removeSet = new BitSet(oldSize);
 		for (int i = 0; i < oldSize; i++) {
 			final D element = this.data.get(i);
-			if (element.removed) {
+			if (element.removed || element.owner != this) {
 				removeSet.set(i);
 				removeCount++;
 			}
